+++ bgstally/colonisation.py
@@ -68,8 +68,8 @@ class Colonisation:
         self.cmdr:str|None = None
 
         # Valid keys for colonisation.json entries. These help avoid sending unnecessary data to third parties or storing unnecessary data in the save file.
-        self.system_keys:list = ['Name', 'StarSystem', 'SystemAddress', 'Claimed', 'Builds', 'Notes', 'Population', 'Economy', 'Security' 'RScync', 'Architect', 'Rev', 'Bodies', 'EDSMUpdated', 'Hidden', 'SpanshUpdated', 'RCSync', 'BuildSlots']
-        self.build_keys:list = ['Name', 'Plan', 'State', 'Base Type', 'Body', 'BodyNum', 'MarketID', 'Track', 'StationEconomy', 'Layout', 'Location', 'BuildID', 'ProjectID', 'TotalCost']
+        self.system_keys:list = ['Name', 'StarSystem', 'SystemAddress', 'Claimed', 'Builds', 'Notes', 'Population', 'Economy', 'Security' 'RScync', 'Architect', 'Rev', 'Bodies', 'EDSMUpdated', 'Hidden', 'SpanshUpdated', 'RCSync', 'BuildSlots', 'RCCommander', 'RCOpen']
+        self.build_keys:list = ['Name', 'Plan', 'State', 'Base Type', 'Body', 'BodyNum', 'MarketID', 'Track', 'StationEconomy', 'Layout', 'Location', 'BuildID', 'ProjectID', 'TotalCost', 'Readonly']
         self.progress_keys:list = ['MarketID', 'Updated', 'ConstructionProgress', 'ConstructionFailed', 'ConstructionComplete', 'ProjectID', 'Required', 'Delivered']
 
         # Load base commodities, types, costs, and saved data
@@ -110,6 +110,8 @@ class Colonisation:
         if entry.get('Type', None) != None: self.station = entry.get('Type')
         if entry.get('BodyType', None) == 'Station': self.station = entry.get('Body')
         if entry.get("StationName", None): self.station = entry.get('StationName')
+        self.station = re.sub(r"^\$EXT_PANEL_ColonisationShip;", "System Colonisation Ship", f"{self.station}").strip()
+
         if cmdr != None: self.cmdr = cmdr
         if self.current_system != None and self.current_system in entry.get('Body', ' '): self.body = self.body_name(self.current_system, entry.get('Body'))
 
@@ -181,7 +183,14 @@ class Colonisation:
                     Debug.logger.warning(f"Invalid ColonisationConstructionDepot event (no system): {entry}")
                     return
                 progress:dict = self.find_or_create_progress(self.market_id)
+                if progress.get('ProjectID', None) != None and entry.get('ProjectID', None) == None:
+                    entry['ProjectID'] = progress.get('ProjectID', None)
                 self.update_progress(self.market_id, entry)
+
+                # Make sure the build is connnected to the project
+                build = self.find_build(system, {'MarketID': self.market_id})
+                if build != None and build.get('ProjectID', None) == None and entry.get('ProjectID', None) != None:
+                    self.modify_build(system, build.get('BuildID', ''), {'ProjectID': entry.get('ProjectID', None)})
 
             case 'Docked':
                 self._update_market(self.market_id)
@@ -191,13 +200,13 @@ class Colonisation:
 
                 # Find and add system and build for construciton sites.
                 # Colonisation ship is always the first build. Construction site can be any build
-                if '$EXT_PANEL_ColonisationShip' in f"{self.station}" or 'Construction Site' in f"{self.station}":
+                if 'Colonisation Ship' in f"{self.station}" or 'Construction Site' in f"{self.station}":
                     Debug.logger.debug(f"Docked at construction site. Finding/creating system and build")
                     if system == None: system = self.find_or_create_system({'StarSystem': self.current_system, 'SystemAddress' : self.system_id})
                     build = self.find_or_create_build(system, {'MarketID': self.market_id, 'Name': self.station, 'Body': self.body})
                     build_state = BuildState.PROGRESS
                 # Complete station so find it and add/update as appropriate.
-                elif system != None and self.station != 'FleetCarrier' and re.search(r"^(...\-...$|\$)", f"{self.station}") == None :
+                elif system != None and self.station not in ['FleetCarrier', 'SquadronCarrier'] and re.search(r"^(...\-...$|\$)", f"{self.station}") == None :
                     Debug.logger.debug(f"Docked at site. Finding/creating system and build {self.market_id} {self.station}")
                     build = self.find_or_create_build(system, {'MarketID': self.market_id, 'Name': self.station, 'Body': self.body})
                     build_state = BuildState.COMPLETE
@@ -402,7 +411,7 @@ class Colonisation:
         ''' Find a system by name or plan, or create it if it doesn't exist '''
         system:dict|None = self.find_system(data)
         if system == None:
-            return self.add_system(data)
+            return self.add_system(data, False, self.bgstally.state.ColonisationRCAPIKey.get() != None)
 
         return system
 
@@ -573,7 +582,7 @@ class Colonisation:
         if data.get('Name', '') == '' or data.get('Name', '') == ' ': data['Name'] = None
 
         # Colonisation ship must be build 0
-        if data.get('Name', None) != None and '$EXT_PANEL_ColonisationShip;' in data.get('Name', '') and len(builds) > 0:
+        if data.get('Name', None) != None and 'System Colonisation Ship' in data.get('Name', '') and len(builds) > 0:
             return builds[0]
 
         # An existing/known build?
@@ -671,6 +680,24 @@ class Colonisation:
 
         return data
 
+    @catch_exceptions
+    def move_build(self, system, row:int, new_row:int) -> None:
+        ''' Move a build to a new position in the list '''
+        if isinstance(system, int): system = self.systems[system]
+
+        if row < 0 or row >= len(system['Builds']) or new_row < 0 or new_row >= len(system['Builds']):
+            Debug.logger.warning(f"Cannot move build - invalid build index: {row} to {new_row}")
+            return
+
+        build:dict = system['Builds'].pop(row)
+        system['Builds'].insert(new_row, build)
+
+        if system.get('RCSync', False) == True and system.get('SystemAddress', None) != None:
+            RavenColonial(self).update_build_order(system)
+
+        self.save('Build moved')
+        self.bgstally.ui.window_colonisation.update_display()
+
 
     @catch_exceptions
     def remove_build(self, system:dict|int, ind:int|str, silent:bool = False) -> None:
@@ -705,7 +732,8 @@ class Colonisation:
             pid:str|None = build.get('ProjectID', None)
             if pid == None:
                 p:dict|None = self.find_progress(build.get('MarketID', ''))
-                pid = p.get('ProjectID', None)
+                if p != None:
+                    pid = p.get('ProjectID', None)
 
             if pid != None:
                 RavenColonial(self).delete_project(pid)
@@ -817,11 +845,11 @@ class Colonisation:
             'State': BuildState.COMPLETE,
             'Track': False,
             'TotalCost': sum(p.get('Required').values()),
+            'Readonly': True,
             'Name': re.sub(r"(\w+ Construction Site:|\$EXT_PANEL_ColonisationShip;|System Colonisation Ship) ", "", build.get('Name', ''))
         }
-        data['MarketID'] = None if '$EXT_PANEL_ColonisationShip;' in build.get('Name', '') else build.get('MarketID', None)
+        data['MarketID'] = None if 'System Colonisation Ship' in build.get('Name', '') else build.get('MarketID', None)
         self.modify_build(system, build.get('BuildID', ''), data)
-
         return True
 
 
@@ -1065,13 +1093,6 @@ class Colonisation:
                     state = BuildState.PROGRESS
             return state.value
 
-        # Builds order
-        def build_order(item:dict) -> int:
-            match self.get_build_state(item):
-                case BuildState.COMPLETE: return 0
-                case BuildState.PROGRESS: return 1
-                case _: return 2
-
         # We sort the order of systems when saving so that in progress systems are first, then planned, then complete.
         # Fortuitously our desired order matches the reverse alpha of the states
         # We also clean up the system and build entires.
@@ -1079,8 +1100,6 @@ class Colonisation:
         for s in list(sorted(self.get_all_systems(), key=sort_order, reverse=True)):
             system:dict = {k: v for k, v in s.items() if k in self.system_keys}
             builds:list = []
-            if len(system['Builds']) > 1:
-                system['Builds'] = [system['Builds'][0]] + list(sorted(system['Builds'][1:], key=build_order))
             for i, b in enumerate(system['Builds']):
                 if i > 0 and b.get('Base Type', '') == '' and b.get('Name', '') == '': continue
                 build:dict = {k: v for k, v in b.items() if k in self.build_keys and v not in ['', "\u0001"]}
@@ -1120,6 +1139,9 @@ class Colonisation:
                     self._generate_buildid(build.get('MarketID', None))
                     self.dirty = True
                 markets += [v for k, v in build.items() if k == 'MarketID' and v != None and v != '']
+                if build.get('Readonly', None) == None:
+                    build['Readonly'] = (build.get('State', BuildState.PLANNED) == BuildState.COMPLETE)
+                    self.dirty = True
 
         for p in dict.get('Progress', []):
             # Clean out old progress entries that are no longer relevant